refactors files to remove the core
